auth sayfaları context yapısına gore revıze edıldı. kayıt ol ve logın sayfaları suan ıcın sorunsuz calısıyor. Note.jsx olusturuldu ve html yapı component yapısına gore bolundu
