+++ src/DialogApp.jsx
@@ -1,6 +1,25 @@
 import { useRef } from "react";
 
 export default function DialogApp() {
+  const [count, setCount] = useState(0);
+  const otherCount = useRef(0);
+
+  function handleClick() {
+    setCount(count + 1);
+  }
+
+  function handleOtherClick() {
+    if (otherCount.current >= 10) {
+      alert('Çok tıkladın yeter');
+      return;
+    }
+  
+
+    otherCount.current++;
+    console.log(otherCount.current);
+  
+  }
+
   const deleteDialogRef = useRef(null);
   const archiveDialogRef = useRef(null);
