--- src/components/ProtectedRoute.jsx
@@ -1,10 +1,0 @@
-import { Navigate } from "react-router-dom";
-import { isAuthenticated } from "../pages/utils/auth";
-
-export default function ProtectedRoute({ children }) {
-  if (!isAuthenticated()) {
-    return <Navigate to="/login" replace />;
-  }
-
-  return children;
-}
