--- src/pages/utils/auth.js
@@ -1,3 +1,0 @@
-export const isAuthenticated = () => {
-  return !!localStorage.getItem("userToken");
-};
